refactor(views): cleanout what we are not going to ship and simplify pre-testing

+++ BlueEagle/Models/HRSample.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct HRSample: Codable {
-    var rate: Int
+    var rate: Double
     var at: Date = Date()
     
     func toJson() -> String {
